requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
CacheControl==0.12.10
cachetools==4.2.4
catkin-pkg==0.4.24
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
cryptography==36.0.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.5
django-braces==1.15.0
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-countries==7.2.1
django-extensions==3.1.5
django-extra-fields==3.0.2
django-filter==21.1
django-heroku==0.3.1
django-jazzmin==2.4.9
django-oauth-toolkit==1.6.3
django-phonenumber-field==6.0.0
django-randompinfield==0.4.7
django-rest-auth==0.9.5
django-rest-authemail==2.1.1
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
django-sendsms==0.2.3
django-staticfiles==1.2.1
django-templated-mail==1.1.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docker==5.0.3
docutils==0.18.1
drf-yasg==1.20.0
facebook-sdk==3.1.0
fcm-django==1.0.5
filelock==3.6.0
firebase-admin==5.1.0
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
googlemaps==4.5.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
jws==0.1.3
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
msgpack==1.0.3
mysqlclient==2.1.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumberslite==8.12.40
Pillow==8.4.0
pipenv==2022.3.28
platformdirs==2.5.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pydotplus==2.0.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.6
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
python-magic==0.4.24
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.3
pywin32==227
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
slugify==0.0.1
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sorl-thumbnail==12.7.0
sqlparse==0.4.2
stripe==2.27.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
websocket-client==1.2.3
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
zope.interface==5.4.0